pyproject.toml
[tool.poetry]
name = "aldtts-core"
version = "0.4.0"
description = "Interventional test for independence"
authors = ["bela127 <bhb127@outlook.de>"]
license = "MIT"
packages = [
    { include = "aldtts/core", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
alts-core = { path = "/home/bela/Cloud/code/Git/alts-core", develop = true }
#distribution-data-generation = {git = "https://github.com/bela127/distribution_data_generation.git"}
#active-learning-ts = { path = "/home/bela/Cloud/code/Git/active_learning_ts", develop = true }
#active-learning-ts = {git = "https://github.com/bela127/active_learning_ts.git"}
libclang = "<12.0"
coloredlogs = "^15.0.1"
ipython = "^8.0.1"
matplotlib = "^3.5.1"
nptyping = ">=2.0.0a1"
scikit-learn = "^1.0.2"
numpy = "^1.22.3"
scipy = "1.8"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
mypy = "^0.910"
black = "^21.6b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[mypy]
plugins = "numpy.typing.mypy_plugin"
